+++ Source/SwiftLintFramework/Configuration/Configuration+Merging.swift
@@ -21,7 +21,7 @@ extension Configuration {
             warningThreshold: mergedWarningTreshold(with: childConfiguration),
             reporter: reporter,
             cachePath: cachePath,
-            allowZeroLintableFiles: childConfiguration.allowZeroLintableFiles,
+            allowZeroLintableFiles: allowZeroLintableFiles || childConfiguration.allowZeroLintableFiles,
             strict: childConfiguration.strict,
             lenient: childConfiguration.lenient,
             baseline: childConfiguration.baseline,